requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
mysqlclient
pandas
sklearn
tensorflow
tensorflow_hub
pytest
mock
pytest-mock
